+++ Sources/StreamChatUI/Composer/ComposerView.swift
@@ -59,6 +59,13 @@ open class ComposerView: _View, ThemeProvider {
         .withoutAutoresizingMaskConstraints
         .withAccessibilityIdentifier(identifier: "sendButton")
 
+    /// A button to record an VoiceRecording.
+    public private(set) lazy var recordButton: RecordButton = components
+        .recordButton
+        .init()
+        .withoutAutoresizingMaskConstraints
+        .withAccessibilityIdentifier(identifier: "recordButton")
+
     /// A view for showing a cooldown when Slow Mode is active.
     public private(set) lazy var cooldownView: CooldownView = components
         .cooldownView.init()
@@ -155,6 +162,9 @@ open class ComposerView: _View, ThemeProvider {
         trailingContainer.distribution = .equal
         trailingContainer.directionalLayoutMargins = .zero
         trailingContainer.addArrangedSubview(sendButton)
+        if components.isVoiceRecordingEnabled {
+            trailingContainer.addArrangedSubview(recordButton)
+        }
         trailingContainer.addArrangedSubview(cooldownView)
         trailingContainer.addArrangedSubview(confirmButton)
         cooldownView.isHidden = true
@@ -176,7 +186,7 @@ open class ComposerView: _View, ThemeProvider {
         titleLabel.centerXAnchor.pin(equalTo: centerXAnchor).isActive = true
         titleLabel.pin(anchors: [.top, .bottom], to: headerView)
 
-        [sendButton, confirmButton]
+        [sendButton, confirmButton, recordButton]
             .forEach { button in
                 button.pin(anchors: [.width], to: 35)
                 button.pin(anchors: [.height], to: 40)